+++ app/components/HowItWorks/HowItWorks.tsx
