+++ app/page.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import Navbar from './components/Navbar';
-import Footer from './components/Footer';
-import TryChatting from './components/TryChatting';
-import CreateYourOwn from './components/CreateYourOwn';
-import HowItWorks from './components/HowItWorks';
-import UsersLoveIt from './components/UsersLoveIt';
-import CustomAgent from './components/CustomAgent';
-import YouCanAlso from './components/YouCanAlso';
+import Navbar from './components/Navbar/Navbar';
+import Footer from './components/Footer/Footer';
+import TryChatting from './components/TryChatting/TryChatting';
+import CreateYourOwn from './components/CreateYourOwn/CreateYourOwn';
+import HowItWorks from './components/HowItWorks/HowItWorks';
+import UsersLoveIt from './components/UsersLoveIt/UsersLoveIt';
+import CustomAgent from './components/CustomAgent/CustomAgent';
+import YouCanAlso from './components/YouCanAlso/YouCanAlso';
 
 export default function Home() {
   return (
